# Conflicts:
#	src/main/java/com/example/webadvanced_backend/controllers/SlideTypeController.java
#	src/main/java/com/example/webadvanced_backend/models/Presentation.java

+++ src/main/java/com/example/webadvanced_backend/controllers/PresentationController.java
@@ -30,9 +30,17 @@ public class PresentationController {
     @Autowired
     SlideRepository slideRepository;
     @Autowired
+    QuestionRepository questionRepository;
+    @Autowired
     ContentMultichoiceRepository multichoiceRepository;
     @Autowired
+    ContentHeadingRepository headingRepository;
+    @Autowired
+    ContentParagraphRepository paragraphRepository;
+    @Autowired
     ContentRepository contentRepository;
+    @Autowired
+    VoteRepository voteRepository;
 
     @GetMapping()
     ResponseEntity<?> getPresentationList(Principal principal){
@@ -48,11 +56,10 @@ public class PresentationController {
     @PostMapping(path = "/add")
     public ResponseEntity<?> createAPresentation(@RequestBody CreatePresentationRequest request, Principal principal){
         try {
-            Instant instant = Instant.ofEpochMilli(Long.parseLong(request.getCreatedTime()));
             Presentation presentation;
             presentation = Presentation.builder().name(request.getPresentationName())
                     .user(accountRepository.findByUsername(principal.getName()))
-                    .createdTime(instant)
+                    .createdTime( request.getCreatedTime())
                     .build();
             return ResponseEntity.ok(presentationRepository.save(presentation));
         }
@@ -66,7 +73,7 @@ public class PresentationController {
         try {
             Optional<Presentation> presentation = presentationRepository.findById(Integer.valueOf(id));
             presentation.get().setName(request.getPresentationName());
-            presentation.get().setModifiedTime(Instant.ofEpochMilli(Long.parseLong(request.getEditTime())));
+            presentation.get().setModifiedTime(request.getEditTime());
             return ResponseEntity.ok(presentationRepository.save(presentation.get()));
         } catch (Exception err) {
             return ResponseEntity.internalServerError().body(err.getMessage());
@@ -84,19 +91,42 @@ public class PresentationController {
             // tim cac content id cua cac slide
             List<Integer> listContentId = new ArrayList<>();
             for (Slide s: listSlide) {
-                if(s.getContent().getSlideType() == 1)
-                    listContentId.add(s.getContent().getId());
+                   listContentId.add(s.getContent().getId());
             }
             //tim cac list content
             List<Content> listContent = contentRepository.findAllById(listContentId);
             //xoa content multichoice
             List<ContentMultichoice> listMultichoice = new ArrayList<>();
+            List<ContentParagraph> listParagraph = new ArrayList<>();
+            List<ContentHeading> listHeading = new ArrayList<>();
             for(Content c: listContent ){
-                listMultichoice.addAll(multichoiceRepository.findByContent(c));
+                if(c.getSlideType() == 1)
+                    listMultichoice.addAll(multichoiceRepository.findByContent(c));
+                else if(c.getSlideType() == 2)
+                    listParagraph.addAll(paragraphRepository.findByContent(c));
+                else if(c.getSlideType() == 3)
+                    listHeading.addAll(headingRepository.findByContent(c));
             }
-            multichoiceRepository.deleteAll(listMultichoice);
+            if(!listMultichoice.isEmpty()){
+                //delete votes of each option
+                List<Vote> listVote = new ArrayList<>();
+                for(ContentMultichoice m : listMultichoice){
+                    listVote.addAll(voteRepository.findByOption(m.getOption()));
+                }
+                voteRepository.deleteAll(listVote);
+                multichoiceRepository.deleteAll(listMultichoice);
+
+            }
+            if(!listParagraph.isEmpty())
+                headingRepository.deleteAll(listHeading);
+            if(!listParagraph.isEmpty())
+                paragraphRepository.deleteAll(listParagraph);
             // xoa cac slide
             slideRepository.deleteAll(listSlide);
+            //xoa question
+            List<Question> listQuestion= questionRepository.findByPresentation(presentation);
+            questionRepository.deleteAll(listQuestion);
+            //
             presentationRepository.delete(presentation);
             return ResponseEntity.ok(presentation);
         }
